Version 1.0
Complete rewrite.  The theme of this rewrite was: quit polluting the mapping namespace, allow for most everything to be customized, remove all personal stuff from the script and write documentation.  All maps have been wrapped into a function and all abbreviations are now prefixed with g:phpLeader so you can change this variable and change the prefix for every abbreviation.  Abbreviations are now created in a more modular fashion so it should be much easier to maintain the script.  I have bumped the version to 1.0-- the structure should *finally* be mostly set so I can now focus on extending the script to cover additional php functions.